+++ request_common.c
@@ -489,7 +489,7 @@ int process_encapsulated(ci_request_t * req, const char *buf)
     start = buf + 13; /*strlen("Encapsulated:")*/
     pos = start;
     end = (char *)start;
-    for (; isspace(*pos) && *pos != '\0'; ++pos);
+    for (; isspace((int)*pos) && *pos != '\0'; ++pos);
     while (*pos != '\0') {
         type = get_encaps_type(pos, &val, &end);
         if (type < 0) /*parse error - return "400 bad request"*/
@@ -502,7 +502,7 @@ int process_encapsulated(ci_request_t * req, const char *buf)
 
         assert(start != end);
         pos = end;            /* points after the value of entity.... */
-        for (; (isspace(*pos) || *pos == ',') && *pos != '\0'; ++pos);
+        for (; (isspace((int)*pos) || *pos == ',') && *pos != '\0'; ++pos);
     }
     req->hasbody = hasbody;
     return EC_100;
